refactor: Move HTTP utility functions to TagHttpUtils and simplify code structure
